feat: new validations and beginning of refactoring of ESN classes
